Remove old viewmodel and use the main one to refresh a folder
